Irritating Gameplay
Removed an irritating piece that required user to hit the button after an invalid input or already visited cell message.  This was causing the user (me) to naturally input the next selection, only to be greeted by another error since the code was waiting for an input first.  Removing the required keypress did not affect the game flow since the message is still shown while waiting for new input.