Issue #12
Added DoubleBuffer = true; and SuspendLayout();/ResumeLayout(); to all instances where the board is drawn.  This creates a noticable improvement in the population of the grid, but the fundamental flaw is that we are drawing 400 buttons at most to the form.  This WILL cause lag in the paint and it is not possible to prevent it.  Labels do not work either.