Fixes Issue #8
Added logic to showgameboard and show cell click events that will determine if a gameboard is present and check for gameMode 1.  If these are true, it will show the gameBoard and call the losing gameMode routine, or find a non-live cell, set hasBeenVisited = true, check wincondition, cascade, and showboard.  It also adds 30 seconds to the player score.